3.3.1 large bug fix package
Streamlining with JADE 2.2 and fixing some artefacts from former measurements, furhter implementing the beta fit for regulized method and intercept = 0 fitting for diffusion coefficient fitting (details see changelog)

+++ ade_dls/analysis/regularized.py
@@ -230,7 +230,7 @@ def nnls_reg(df, name, nnls_reg_params, plot_number):
     # Create Tikhonov regularization matrix (2nd derivative)
     def create_tikhonov_matrix(n):
         #create a sparse matrix for computational efficiency
-        D2 = sparse.diags([1, -2, 1], [-1, 0, 1], shape=(n-2, n)).toarray()
+        D2 = sparse.diags([1, -2, 1], [-1, 0, 1], shape=(n-2, n), dtype=float).toarray()
         return D2
     
     #get the length of decay_times
@@ -599,7 +599,7 @@ def nnls_reg_simple(df, name, nnls_reg_params):
 
     #create Tikhonov regularization matrix (second derivative)
     def create_tikhonov_matrix(n):
-        D2 = sparse.diags([1, -2, 1], [-1, 0, 1], shape=(n-2, n)).toarray()
+        D2 = sparse.diags([1, -2, 1], [-1, 0, 1], shape=(n-2, n), dtype=float).toarray()
         return D2
     
     #get the length of decay_times
